Dataset refactoring
Extracted the code for dataset table handler
