+++ backend/dataall/modules/datasets/handlers/__init__.py
@@ -2,7 +2,10 @@
 Contains code with the handlers that are need for async
 processing in a separate lambda function
 """
-from dataall.modules.datasets.handlers import glue_column_handler
+from dataall.modules.datasets.handlers import (
+    glue_column_handler,
+    glue_table_handler
+)
 
-__all__ = ["glue_column_handler"]
+__all__ = ["glue_column_handler", "glue_table_handler"]
